Refactors shared contracts and adds court retrieval
Moves shared DTOs and requests to a dedicated Shared-Contracts project to promote code reuse and consistency across microservices.

Implements court retrieval functionality in the Booking Microservice by adding a PaddleCourtClient that communicates with the Paddle Court Microservice to fetch paddle court data.
This enables the Booking Microservice to display available courts to users when making bookings.